Fixed ApacheResponse.getResponseBodyAsStream() to return all the bytes in the body
ApacheResponse.getResponseBodyAsStream() was returning an InputStream
that only provided bytes from the first body part.

Pulled JDKResponse's ByteArrayCollectionInputStream into its own class,
HttpResponseBodyPartsInputStream, so it could be shared with
ApacheResponse.
